% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_boxplot.R
\name{alpha_boxplot}
\alias{alpha_boxplot}
\title{Making alpha-diversity boxplots}
\usage{
alpha_boxplot(pseq, x_grp, a_index, extra_var = NULL, ...)
}
\arguments{
\item{pseq}{Your phyloseq object.}

\item{x_grp}{The variable of the groups you want to compare.}

\item{a_index}{The alpha-diversity you want to compare.}

\item{extra_var}{Optional: extra variables that you want to include in the comparison.}

\item{...}{Other arguments passed on to microbiome::boxplot_alpha.}
}
\value{
A boxplot with alpha-diversities per group
}
\description{
Making alpha-diversity boxplots
}
\examples{
alpha_boxplot(ps, x_grp = "Source", "Shannon", extra_var = "Time_group")
}
