% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc2.R
\docType{data}
\name{ancombc2_wrappers}
\alias{ancombc2_wrappers}
\alias{ancombc2_params}
\title{ANCOMBC2 Wrappers}
\format{
An object of class \code{list} of length 20.
}
\usage{
ancombc2_params
}
\description{
Functions for running ANCOM-BC2 analyses and displaying results.
}
\section{Functions}{

- `run_ancombc2()`: Runs ANCOM-BC2 analysis on microbiome data.
- `display_ancombc2_results()`: Displays the results in an optionally interactive table.
}

\keyword{datasets}
