% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_abundance.R
\name{barplot_abundance}
\alias{barplot_abundance}
\title{Make an abundance barplot}
\usage{
barplot_abundance(pseq, transf = "compositional", ...)
}
\arguments{
\item{pseq}{Your phyloseq object.}

\item{transf}{Transformation applied to otu counts. Default (and recommended) is relative (aka percentage / "compositional")}

\item{...}{Other arguments to be passed on to both micrbiome::aggregate_rare AND to microbiome::plot_composition.}
}
\value{
An abundance barplot.
}
\description{
Make an abundance barplot
}
\examples{
barplot_abundance(ps, transf = "compositional", prevalence = 5, detection = 0.2, include.lowest = TRUE, sample.sort = NULL, otu.sort = NULL, average_by = NULL, group_by = NULL)
}
