% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_taxonomy.R
\name{clean_taxonomy}
\alias{clean_taxonomy}
\title{Cleaning of taxonomy.}
\usage{
clean_taxonomy(pseq)
}
\arguments{
\item{pseq}{Your phyloseq object.}
}
\value{
Your phyloseq object with cleaned taxonomy table.
}
\description{
Cleaning of taxonomy.
}
\examples{
ps_clean <- clean_taxonomy(ps)
}
