% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Importing data}
\usage{
import_data(name_biom_file = NULL, name_mapping_file = NULL, factors = NULL)
}
\arguments{
\item{name_biom_file}{The name of the biom file.}

\item{name_mapping_file}{The name of the mapping file.}

\item{factors}{A `char` vector of variable names in the metadata to be used as factors.}
}
\value{
pseq A phyloseq object.
}
\description{
This function will import the data you've provided in the config file and make it into a phyloseq object.
In case no config file was provided, test data is loaded.
}
\examples{
import_data(my_biom_file, my_mapping_file)
}
