% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a heatmap}
\usage{
make_heatmap(pseq, transf = "clr", pseudocount = 1, ...)
}
\arguments{
\item{pseq}{Your phyloseq object.}

\item{transf}{What transformation you want to use. Note that a log transformation is highly recommended.}

\item{pseudocount}{What pseudocount to use before log-transformation.}

\item{...}{Other arguments to pass to microbiome::core. Such as prevalence and detection.}
}
\value{
A heatmap plot
}
\description{
Make a heatmap
}
\examples{
make_heatmap(ps, transf = "clr", pseudocount = 1, prevalence = 5, detection = 0.2, include.lowest = TRUE)
}
