% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pca.R
\name{make_pca}
\alias{make_pca}
\title{Principal Component Analysis}
\usage{
make_pca(
  pseq,
  transf = "clr",
  colour_by = NULL,
  ellipse = FALSE,
  longi_lines = NULL,
  pseudocount = 1,
  ...
)
}
\arguments{
\item{pseq}{Your phyloseq object}

\item{transf}{What transformation to apply to the data. Default is CLR, with shift of 1.}

\item{colour_by}{What variable to colour the points by. Default no colouring.}

\item{ellipse}{Whether to include an ellips (t-distribution), TRUE or FALSE.}

\item{longi_lines}{Whether you want to include lines to connect repeated measures, TRUE or FALSE.}

\item{pseudocount}{What pseudocount you want to use in case of (any) log-transformation.}

\item{...}{Additional parameters to be passed on to autoplot.}
}
\value{
A PCA plot
}
\description{
Principal Component Analysis
}
\examples{
make_PCA(ps, "clr", pseudocount = 0.5)
}
