% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_curve.R
\name{rare_curve}
\alias{rare_curve}
\title{Rarecurve for estimating rarefaction depth}
\usage{
rare_curve(pseq, ...)
}
\arguments{
\item{pseq}{Your phyloseq object.}

\item{...}{Other arguments to be passed to vegan::rarecurve}
}
\value{
A rarecurve.
}
\description{
Rarecurve for estimating rarefaction depth
}
\examples{
rare_curve(ps)
}
