% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{rarefy}
\alias{rarefy}
\title{Rarefy otu counts}
\usage{
rarefy(pseq, depth = 0, mc = 1000)
}
\arguments{
\item{pseq}{Your phyloseq object.}

\item{depth}{The depth you want to rarefy to.}

\item{mc}{The minimum count that a sample should have to not be thrown out.}
}
\value{
A phyloseq object where samples rarefied or samples below minimum count are thrown out.
}
\description{
Rarefy otu counts
}
\examples{
rarefy(ps, depth = 4000, mc = 1000)
}
