% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_single.R
\name{rarefy_single}
\alias{rarefy_single}
\title{Rarefy OTU counts}
\usage{
rarefy_single(pseq, depth = 0, mc = 1000)
}
\arguments{
\item{pseq}{Your phyloseq object.}

\item{depth}{The depth you want to rarefy to.}

\item{mc}{The minimum count that a sample should have to not be thrown out.}
}
\value{
A phyloseq object where samples rarefied or samples below minimum count are thrown out.
}
\description{
Rarefy OTU counts
}
\examples{
rarefy_single(ps, depth = 4000, mc = 1000)
}
