% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc2.R
\name{run_ancombc2}
\alias{run_ancombc2}
\title{Run ANCOM-BC2 Analysis}
\usage{
run_ancombc2(ps, params)
}
\arguments{
\item{ps}{A `phyloseq` object containing the microbiome data.}

\item{params}{A named list of parameters for the ANCOM-BC2 analysis. See Details for required parameters.}
}
\value{
A list containing the ANCOM-BC2 analysis results:
\itemize{
  \item \code{res_global}: Results from global tests.
  \item \code{res_pair}: Results from pairwise comparisons.
  \item \code{dunn}: Results from Dunnett's test.
  \item \code{res_trend}: Results from trend analysis.
  \item \code{res}: Primary analysis results.
}
}
\description{
This function executes the ANCOM-BC2 analysis on a given phyloseq object using specified parameters.
ANCOM-BC2 is a differential abundance analysis method for microbiome data.
}
\details{
The `params` list should contain the following elements:
\describe{
  \item{taxonomy_level}{Taxonomic level to be tested (e.g., "Genus").}
  \item{prevalence_cutoff}{Prevalence cut-off for filtering taxa.}
  \item{library_size_cutoff}{Library size cut-off for filtering samples.}
  \item{structural_zero}{Logical, whether to detect structural zeros.}
  \item{lower_bound}{Logical, whether to classify a taxon as a structural zero using its asymptotic lower bound.}
  \item{fixed_terms}{Formula specifying fixed effects.}
  \item{random_terms}{Formula specifying random effects.}
  \item{grouping_variable}{Grouping variable for multi-group tests.}
  \item{p_adj_method}{Method for p-value adjustment (e.g., "holm").}
  \item{alpha}{Significance level.}
  \item{global}{Logical, whether to perform a global test.}
  \item{pairwise}{Logical, whether to perform pairwise tests.}
  \item{dunnet}{Logical, whether to perform Dunnett's test.}
  \item{trend}{Logical, whether to perform a trend test.}
  \item{pseudo_sens}{Logical, whether to perform pseudocount sensitivity analysis.}
  \item{s0_perc}{Percentile of standard error values for sensitivity analysis.}
  \item{n_cores}{Number of CPU cores to use.}
  \item{verbose}{Logical, whether to print verbose output.}
  \item{iter}{Number of iterations for REML and EM algorithms.}
  \item{bootstrap}{Number of bootstrap samples.}
}
}
\examples{
\dontrun{
result <- run_ancombc2(ps, params)
}

}
