% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc2.R
\name{ancombc2_wrappers}
\alias{ancombc2_wrappers}
\title{ANCOMBC2 Wrappers}
\description{
Functions for running ANCOM-BC2 analyses and displaying results.
}
\section{Functions}{

- `run_ancombc2()`: Runs ANCOM-BC2 analysis on microbiome data.
- `display_results()`: Displays the results in an interactive table.
}

