% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc2.R
\name{display_ancombc2_results}
\alias{display_ancombc2_results}
\title{Display ANCOM-BC2 Results}
\usage{
display_ancombc2_results(out, selected_results = NULL)
}
\arguments{
\item{out}{The output object from `run_ancombc2` (ANCOMBC::ancombc2 result).}

\item{selected_results}{A character vector specifying which results to display.
Possible values include "global", "pairwise", "dunnett", and "trend".
If NULL, the primary analysis results are displayed.}
}
\value{
A `DT::datatable` object displaying the selected results.
}
\description{
This function presents the results of an ANCOM-BC2 analysis in an interactive table.
}
\examples{
\dontrun{
display_results(result, selected_results = c("global", "pairwise"))
}

}
