% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import your data}
\usage{
import_data(name_biom_file, name_mapping_file)
}
\arguments{
\item{name_biom_file}{the biom file, in the form of "config$biom"}

\item{name_mapping_file}{the mapping file, in the form of "config$mapping"}
}
\value{
a phyloseq object
}
\description{
Give your biom file and mapping file to create a phyloseq object.
If the configuration file is not loaded, then a test object will be made.
}
