% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_loop_functions.R
\name{rarefy_multiple}
\alias{rarefy_multiple}
\alias{calculate_alpha_df}
\title{Subsample-loop}
\usage{
rarefy_multiple(
  pseq,
  sample.size = NULL,
  min_size_threshold = NULL,
  iter = 10,
  replace = FALSE,
  seeds = NULL,
  ...
)

calculate_alpha_df(pseq_list, measures = NULL)
}
\arguments{
\item{pseq}{Your phyloseq object (or OTU-table).}

\item{sample.size}{Your preferred sample size for rarefying.}

\item{min_size_threshold}{Minimum count of otu for a sample to be included.}

\item{iter}{The amount of times you would like to rarefy (amount of comparisons).}

\item{replace}{Perform subsampling with replacement or without (TRUE or FALSE).}

\item{seeds}{Vector of random seeds that you want to use. Possible to leave empty and have seed range to be 1:iterations.}

\item{...}{Additional arguments to  be passed on to phyloseq::rarefy_even_depth}

\item{pseq_list}{The list of phyloseq objects received from rarefy_multiple function.}

\item{measures}{The alpha-diversity measures that you want to compute.}
}
\value{
A list of phyloseq objects with length of amount of iterations.

A dataframe giving the alpha-diversities requested for each phyloseq object.
}
\description{
Subsample-loop
}
\section{Functions}{
\itemize{
\item \code{calculate_alpha_df()}: Calculating of alpha-diversities

}}
\examples{
ps_list <- rarefy_multiple(ps, sample.size = 8000, iter = 100)
alpha_df <- calculate_alpha_df(ps_list, measures = c("Shannon", "Simpson", "Chao1"))
}
