% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_loop_functions.R
\name{rarefy_multiple}
\alias{rarefy_multiple}
\alias{calculate_alpha_df}
\alias{calculate_average_alpha_ps}
\alias{multiple_test_alpha}
\alias{multiple_permanova}
\alias{permanova_average}
\alias{acat}
\title{Subsample-loop}
\usage{
rarefy_multiple(
  pseq,
  sample.size = NULL,
  iter = 10,
  replace = FALSE,
  seeds = NULL,
  ...
)

calculate_alpha_df(pseq_list, measures = NULL)

calculate_average_alpha_ps(
  alpha_dataframe,
  alpha_div = NULL,
  averagef = "median"
)

multiple_test_alpha(
  alpha_dataframe,
  pseq,
  alpha_div,
  variable,
  method = "wilcox.test",
  pair_by = NULL
)

multiple_permanova(
  list_of_ps,
  distance,
  variable,
  permutations,
  pseudocount = 1,
  ps_ref = NULL,
  longit = NULL
)

permanova_average(results_adonis, plot = TRUE)

acat(p_values)
}
\arguments{
\item{pseq}{Your original pseq object. This is required only for the metadata.}

\item{sample.size}{Your preferred sample size for rarefying.}

\item{iter}{The amount of times you would like to rarefy (amount of comparisons).}

\item{replace}{Perform subsampling with replacement or without (TRUE or FALSE).}

\item{seeds}{Vector of random seeds that you want to use. Possible to leave empty and have seed range to be 1:iterations.}

\item{...}{Additional arguments to  be passed on to phyloseq::rarefy_even_depth}

\item{pseq_list}{(2) The list of phyloseq objects received from rarefy_multiple function.}

\item{measures}{The alpha-diversity measures that you want to compute.}

\item{alpha_dataframe}{(4) The dataframe with alpha-diversities received from calculate_alpha_df function.}

\item{alpha_div}{The alpha-diversity measures that you want to test.}

\item{averagef}{The type of average function you want to use: median, mean, min (minimum) or max (maximum). Default is median (also recommended).}

\item{variable}{Character. Metadata variable(s) to test (e.g., "Group" or "Var1 + Var2").}

\item{method}{What test should be performed.}

\item{pair_by}{In case of paired analysis, what is variable is the data paired on.}

\item{list_of_ps}{List of rarefied phyloseq objects (from \code{rarefy_multiple}).}

\item{distance}{Character. Distance metric to use (e.g., "bray", "aitchison").}

\item{permutations}{Integer. Number of permutations for PERMANOVA.}

\item{pseudocount}{Numeric. Pseudocount to add to OTU table for Aitchison distance (default: 1).}

\item{ps_ref}{Optional. Reference phyloseq object for metadata subsetting.}

\item{longit}{Optional. Name of metadata variable for paired/longitudinal analysis.}

\item{results_adonis}{List of PERMANOVA results as returned by \code{multiple_permanova}.}

\item{plot}{Logical; if \code{TRUE}, boxplots of pseudo-F and p-values are generated. Default is \code{TRUE}.}

\item{p_values}{Numeric vector of p-values to combine. All values must be strictly between 0 and 1.}
}
\value{
(1) A list of phyloseq objects with length of amount of iterations.

(2) A dataframe giving the alpha-diversities requested for each phyloseq object.

(3) A dataframe with the average alpha-diversities.

(4) A vector with the p-values for each phyloseq object.

A list of adonis2 results, one per rarefied dataset.

A named list with:
\item{results}{A data frame containing the median and IQR of pseudo-F statistics, and the aggregated Cauchy p-value.}
Additional attributes:
\item{pseudoF_values}{Vector of pseudo-F statistics from each rarefaction.}
\item{p_values}{Vector of p-values from each rarefaction.}

A single aggregated p-value (numeric).
}
\description{
These functions are part of a method where multiple different subsamples are taken from a phyloseq object,
to give an idea as to what the spread is subsampled sets.
Because these functions are highly dependent on each other they are put together in one description. Different functions are numbered in the description.

Performs PERMANOVA (adonis2) on each rarefied phyloseq object to test for group differences in community composition.

Aggregates the pseudo-F statistics and p-values from a list of PERMANOVA results (as produced by \code{multiple_permanova}). Returns summary statistics and, optionally, visualizes the distributions.

Combines multiple p-values into a single aggregated p-value using the Aggregated Cauchy Association Test (ACAT).
}
\section{Functions}{
\itemize{
\item \code{rarefy_multiple()}: Making multiple subsampled phyloseq objects

\item \code{calculate_alpha_df()}: Calculating of alpha-diversities

\item \code{calculate_average_alpha_ps()}: Calculating of the averages of alpha-diversities

\item \code{multiple_test_alpha()}: Testing of alpha-diversities between groups

\item \code{multiple_permanova()}: Run PERMANOVA on Multiple Rarefied Phyloseq Objects

\item \code{permanova_average()}: Aggregate PERMANOVA Results Across Multiple Rarefactions

\item \code{acat()}: #' Aggregated Cauchy Association Test (ACAT) for P-value Combination

}}
\examples{
ps_list <- rarefy_multiple(ps, sample.size = 8000, iter = 100)
alpha_df <- calculate_alpha_df(ps_list, measures = c("Shannon", "Simpson", "Chao1"))
average_alphas <- calculate_average_alpha_ps(alpha_df, alpha_div = c("Shannon", "Simpson"), averagef = "min")
alpha_p_values <- multiple_test_alpha(alpha_df, pseq, alpha_div = "Shannon", variable = "HealthStatus", method = "wilcoxon.test", pair_by = "SubjectID")
median(alpha_p_values)
permanova_results <- multiple_permanova(ps_list, distance = "aitchison", variable = "Source", permutations = 9999, longit = "SubjectID")
results <- permanova_average(results_adonis)
combined_p <- acat(c(0.01, 0.03, 0.2))
}
